Fixed review comments
