+++ cmd/ocm-support/create/role_bindings/organization_role_binding/cmd.go
@@ -9,6 +9,7 @@ import (
 	"github.com/openshift-online/ocm-support-cli/cmd/ocm-support/utils"
 	"github.com/openshift-online/ocm-support-cli/pkg/account"
 	"github.com/openshift-online/ocm-support-cli/pkg/organization"
+	"github.com/openshift-online/ocm-support-cli/pkg/role"
 	rolebinding "github.com/openshift-online/ocm-support-cli/pkg/role_binding"
 )
 
@@ -31,17 +32,17 @@ var CmdCreateOrganizationRoleBinding = &cobra.Command{
 		// validates the account
 		err = account.ValidateAccount(accountID, connection)
 		if err != nil {
-			return fmt.Errorf("%v", err)
+			return err
 		}
 		// validates the organization
 		err = organization.ValidateOrganization(orgID, connection)
 		if err != nil {
-			return fmt.Errorf("%v", err)
+			return err
 		}
-		// validates the role binding
-		err = rolebinding.ValidateRole(roleID, connection)
+		// validates the role
+		err = role.ValidateRole(roleID, connection)
 		if err != nil {
-			return fmt.Errorf("%v", err)
+			return err
 		}
 		return nil
 	},
@@ -57,7 +58,7 @@ func runCreateOrganizationRoleBinding(cmd *cobra.Command, argv []string) error {
 	}
 	rb, err := rolebinding.AddRoleBinding(accountID, roleID, rolebinding.OrganizationRoleBinding, &orgID, connection)
 	if err != nil {
-		return fmt.Errorf("failed to validate role %s: %v", roleID, err)
+		return fmt.Errorf("failed to add role binding for role %s: %v", roleID, err)
 	}
 	utils.PrettyPrint(rolebinding.PresentRoleBinding(rb))
 	return nil
