+++ cmd/ocm-support/get/accounts/cmd.go
@@ -10,7 +10,7 @@ import (
 	"github.com/openshift-online/ocm-support-cli/cmd/ocm-support/utils"
 	"github.com/openshift-online/ocm-support-cli/pkg/account"
 	"github.com/openshift-online/ocm-support-cli/pkg/registry_credential"
-	"github.com/openshift-online/ocm-support-cli/pkg/role"
+	rolebinding "github.com/openshift-online/ocm-support-cli/pkg/role_binding"
 )
 
 var args struct {
@@ -103,7 +103,7 @@ func run(cmd *cobra.Command, argv []string) error {
 
 		var roles []*v1.RoleBinding
 		if args.fetchRoles {
-			roles, err = role.GetAccountRoles(acc.ID(), connection)
+			roles, err = rolebinding.GetAccountRoleBindings(acc.ID(), connection)
 			if err != nil {
 				return fmt.Errorf("failed to fetch roles: %s", err)
 			}
