+++ pkg/role_binding/role_binding.go
@@ -7,7 +7,6 @@ import (
 
 	sdk "github.com/openshift-online/ocm-sdk-go"
 	v1 "github.com/openshift-online/ocm-sdk-go/accountsmgmt/v1"
-	"github.com/openshift-online/ocm-support-cli/pkg/role"
 	"github.com/openshift-online/ocm-support-cli/pkg/types"
 )
 
@@ -17,8 +16,8 @@ type RoleBinding struct {
 	HREF           string
 	AccountID      string
 	RoleID         string
-	OrganizationID string
-	SubscriptionID string
+	OrganizationID string `json:",omitempty"`
+	SubscriptionID string `json:",omitempty"`
 	CreatedAt      time.Time
 	UpdatedAt      time.Time
 	Type           string
@@ -29,22 +28,6 @@ const (
 	OrganizationRoleBinding = "Organization"
 	ApplicationRoleBinding  = "Application"
 )
-
-func ValidateRole(roleID string, conn *sdk.Connection) error {
-	availableRoles, err := role.GetRoles(conn)
-
-	if err != nil {
-		return fmt.Errorf("can't validate role binding : %v", err)
-	}
-
-	for _, avavailableRole := range availableRoles {
-		if avavailableRole.ID() == roleID {
-			return nil
-		}
-	}
-
-	return fmt.Errorf("role not found")
-}
 
 func CreateRoleBinding(accountID string, roleID string, roleType string, resourceID *string) (*v1.RoleBinding, error) {
 	newRoleBinding := v1.NewRoleBinding().AccountID(accountID).RoleID(roleID).Type(roleType)
@@ -120,4 +103,25 @@ func DeleteRoleBinding(accountID string, roleBindingKey string, roleType string,
 		return fmt.Errorf("can't delete role binding : %v", err)
 	}
 	return nil
+}
+
+func GetAccountRoleBindings(accountID string, conn *sdk.Connection) ([]*v1.RoleBinding, error) {
+	query := fmt.Sprintf("account_id = '%s'", accountID)
+	response, err := conn.AccountsMgmt().V1().RoleBindings().List().
+		Parameter("search", query).
+		Send()
+
+	if err != nil {
+		return nil, fmt.Errorf("can't retrieve roles for account %s : %v", accountID, err)
+	}
+
+	return response.Items().Slice(), nil
+}
+
+func PresentRoleBindings(roleBindings []*v1.RoleBinding) []string {
+	var roleList []string
+	for _, roleBinding := range roleBindings {
+		roleList = append(roleList, roleBinding.Role().ID())
+	}
+	return roleList
 }
